requirements.txt
Flask==0.8
Flask-SQLAlchemy==0.15
Jinja2==2.6
SQLAlchemy==0.7.5
Scrapy==0.14.1
Werkzeug==0.8.3
certifi==0.0.8
chardet==1.0.1
distribute==0.6.24
lxml==2.3.3
psycopg2==2.4.4
readability-lxml==0.2.3
requests==0.10.4
simplejson==2.3.2
wsgiref==0.1.2
